backend/requirements.txt
asgiref==3.5.0
Django==4.0.4
django-appconf==1.0.5
django-cors-headers==3.12.0
django-filter==21.1
django-js-asset==2.0.0
django-measurement==3.2.4
django-mptt==0.13.4
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Markdown==3.3.6
measurement==3.2.0
mpmath==1.2.1
phonenumbers==8.12.47
Pillow==9.1.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
sympy==1.10.1
tzdata==2022.1
